fixed type mismatch in smart contract example

+++ smart_contract_examples/smart_escrow_example.c
@@ -3,7 +3,7 @@
 
 typedef struct PACKED
 {
-    uint64_t amount;
+    uint32_t amount;
     addr_t recipient;
 } params1_t;
